+++ src/options.tsx
@@ -2,9 +2,10 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import OptionsUI from './options-ui.tsx'
 import { initI18n } from './languages.ts'
+import { initKeymap } from './keymapstore.ts'
 import './css/options.css'
 
-initI18n().then(() => {
+Promise.all([initI18n(), initKeymap()]).then(() => {
   ReactDOM.createRoot(document.getElementById('root')!).render(
     <React.StrictMode>
       <OptionsUI />